pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alerts"
version = "0.1.0"
description = "SMARTS Alert False Positive Analyzer - LLM-powered compliance filter for surveillance alerts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Alerts Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "a2a-sdk[http]>=0.2.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
alerts = "alerts.main:main"
alerts-insider-trading-server = "alerts.a2a.insider_trading_server:main"
alerts-wash-trade-server = "alerts.a2a.wash_trade_server:main"
alerts-orchestrator-server = "alerts.a2a.orchestrator_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
alerts = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
